+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 Improvements:
 
 * Improve the inheritance of code and avoid duplicate code.
+* Profiles: notify profile and/or admin when profile is added.
+* Profiles: notify profile when profile is edited.
 
 Bugfixes: